Reorganize VMC functionality into separate directory
