+++ src/data/vmc/jobs.js
@@ -1,4 +1,4 @@
-import { fetch } from '../data/fetch';
+import { fetch } from '../fetch';
 
 export const listJobs = async domain => {
   const config = domain ? { headers: { 'X-Tenant-Id': domain } } : {};
