+++ src/gcp/detail.js
