+++ src/vmc/organizations/detail.js
@@ -13,11 +13,11 @@ import {
   Page,
   Pill,
 } from '@janus.team/janus-particles';
-import { RequestStatusPanel } from '../common/RequestStatusPanel';
-import { usePoller } from '../data/poller';
-import { getOrganization } from '../data/organizations';
-import { listSDDCs } from '../data/sddcs';
-import { isGossEnabled } from '../config';
+import { RequestStatusPanel } from '../../common/RequestStatusPanel';
+import { usePoller } from '../../data/poller';
+import { getOrganization } from '../../data/vmc/organizations';
+import { listSDDCs } from '../../data/vmc/sddcs';
+import { isGossEnabled } from '../../config';
 
 const SDDCCard = ({ organization, sddc, domain }) => {
   const card_name_span = <span data-testid="sddc-card-name">{sddc.name}</span>;
