+++ src/vmc/sddcs/detail.js
@@ -3,15 +3,15 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { Breadcrumbs, Button, Collection, Page } from '@janus.team/janus-particles';
-import { RequestStatusPanel } from '../common/RequestStatusPanel';
-import { usePoller } from '../data/poller';
-import { listJobs } from '../data/jobs';
-import { getOrganization } from '../data/organizations';
-import { getSDDC } from '../data/sddcs';
-import { listVMs } from '../data/vms';
+import { RequestStatusPanel } from '../../common/RequestStatusPanel';
+import { usePoller } from '../../data/poller';
+import { listJobs } from '../../data/vmc/jobs';
+import { getOrganization } from '../../data/vmc/organizations';
+import { getSDDC } from '../../data/vmc/sddcs';
+import { listVMs } from '../../data/vmc/vms';
 import { GOSSEnrollmentSection } from './goss';
 import { GOSSSettingsSection } from './gossSettings';
-import '../gossSettings.css';
+import '../../gossSettings.css';
 
 import HelixUI from 'helix-ui/dist/js/helix-ui.min';
 HelixUI.initialize();
