+++ src/vmc/sddcs/detail.test.js
